+++ lnmp-docker.ps1
@@ -315,7 +315,7 @@ switch($first){
     }
 
     httpd-config {
-      wsl -d $DistributionName lnmp-docker.sh httpd-config $other
+      wsl -d $DistributionName lnmp-docker httpd-config $other
     }
 
     backup {
@@ -361,15 +361,15 @@ switch($first){
     }
 
     dashboard {
-      wsl -d $DistributionName lnmp-docker.sh dashboard
+      wsl -d $DistributionName lnmp-docker dashboard
     }
 
     new {
-      wsl -d $DistributionName lnmp-docker.sh new $other
+      wsl -d $DistributionName lnmp-docker new $other
     }
 
     nginx-config {
-      wsl -d $DistributionName lnmp-docker.sh nginx-config $other
+      wsl -d $DistributionName lnmp-docker nginx-config $other
     }
 
     swarm-config {
@@ -572,7 +572,7 @@ switch($first){
     }
 
     update-version {
-      wsl -d $DistributionName lnmp-docker.sh update-version
+      wsl -d $DistributionName lnmp-docker update-version
     }
 
     debug {